Breaks up the rest of the system into components
Still lots of tests to be written but the system is now broken
into individual components. This has made the index.js file much
cleaner and will make the whole thing more maintainable in the
future. Tests are all passing and code coverage is marginal.

Some of the tests are pretty dumb and there is a lot better testing
still in need of being written.
