+++ components/off/index.test.js
@@ -1,0 +1,13 @@
+var assert = require('chai').assert
+    , subject = require('./index')
+    , eventStore = require('../store');
+
+describe('off::tests', function () {
+    'use strict';
+
+    it('should have  be a function', function () {
+        assert.isFunction(subject);
+    });
+
+    it('should do things');
+});
